+++ swanlab/utils/__init__.py
@@ -1,4 +1,6 @@
+"""
+工具函数，用于提供一些常用的工具函数，与项目本身无关，在项目任何文件中都可以使用
+"""
 from .font import generate_color, FONT, COLOR_LIST
 from .time import create_time
 from .file import get_a_lock, check_load_json_yaml
-from .package import get_package_version, version_limit
