+++ test/unit/auth/pytest_login.py
@@ -8,8 +8,7 @@ r"""
     测试登录
 """
 from swanlab.auth.login import login_by_key
-from tutils.config import nanoid, CONFIG, PACKAGE_PATH
-from swanlab.utils.package import get_host_api
+from tutils.config import CONFIG
 import pytest
 
 
@@ -20,3 +19,5 @@ async def test_login_success():
     """
     login_info = await login_by_key(CONFIG.get('api-key'))
     assert not login_info.is_fail
+    assert login_info.api_key == CONFIG.get('api-key')
+    assert login_info.__str__() == "OK"
